Add transactions api helper methods.
